+++ src/commands/Destroy.js
@@ -3,10 +3,12 @@ import Command from '../Command'
 import depurar from 'depurar'; const debug = depurar('frey')
 import chalk from 'chalk'
 import _ from 'lodash'
+import Shell from '../Shell'
 
 class Destroy extends Command {
   constructor (name, runtime) {
     super(name, runtime)
+    this.shell = new Shell(runtime)
     this.boot = [
       '_gatherTerraformArgs'
     ]
@@ -37,7 +39,7 @@ class Destroy extends Command {
     cmd = cmd.concat(this.bootCargo._gatherTerraformArgs)
     cmd = cmd.join(' ')
 
-    return this._cmdYesNo(cmd, (err) => {
+    return this.shell._cmdYesNo(cmd, (err) => {
       if (err) {
         return cb(err)
       }