+++ test/test-Shell.js
@@ -1,22 +1,19 @@
-import Command from '../src/Command'
+import Shell from '../src/Shell'
 import { expect } from 'chai'
 
-describe('Command', () => {
+describe('Shell', () => {
   describe('_buildChildEnv', () => {
     it('should convert tf vars', done => {
-      const command = new Command('prepare', {
-        projectDir: '{{{init.os.cwd}}}/frey/production'
-      })
-
-      command.runtime = {
+      const runtime = {
         init: {
           env: {
             FREY_SOMETHING: 'foobar'
           }
         }
       }
+      const shell = new Shell(runtime)
 
-      const env = command._buildChildEnv({'EXTRA': 'Yes Please'})
+      const env = shell._buildChildEnv({'EXTRA': 'Yes Please'})
       expect(env).to.deep.equal({
         'EXTRA': 'Yes Please',
         'FREY_SOMETHING': 'foobar',